task: Unref taskui from the main thread
Because TaskUI-s will deal with UI stuff, they will probably deal with
UI stuff, e.g. widgets. If we were to unref a taksui from a thread, it
would then have to ensure, during its finalize, that any calls such as
gtk_widget_destroy() was done back in the main thread.

To make things easier from TaskUI-s, and because usually the one ref
on takuis will be from their task, let's ensure (and set as a rule) that
unref-ing a taskui must always be done from main thread.