statusbar: Fix warning status-changed w/ no match
There could be warnign "signal 'status-changed' ... found no match" when
the statusbar was connect to a provider for one area/is, but received a
change for another one.

This happens because for :active/:focused donna only switches the
provider for the area, leaving the old one active. The point being that
when switching again (and one might often switch focus between tree &
list, even if focus doesn't stay on tree or long) it switches back,
instead of having to free/alloc each time.

But when the statusbar has the same provider for different areas, this
could happen.

To solve it, the signal "status-changed" is now detailled with the id of
the area.