treeview: Clicks are new defined under [click_modes]
A new treeview option "click_mode" allows to set the click mode for
the treeview. All click options are looked for in the corresponding
click mode, thus allowing to centralize everything, making it (much)
easier to handle, share, or switch from one click mode to another.

Click modes have an option "fallback" to indicate which click mode to go
look into when a click isn't defined. This only works once (i.e. when
looking in a click mode as fallback, there won't be no more falling
back).

In each click_mode there can also be a category "columns" and then
categories for per-column click definitions. Under those, as well as in
the click mode, can be a category "selected" for when the clicked row is
selected.

The paths looked for are, in order:
- clickmode/columns/NAME/selected
- fallback/columns/NAME/selected
- clickmode/selected
- fallback/selected
- clickmode/columns/NAME
- fallback/columns/NAME
- clickmode
- fallback

Tree visuals "clicks" have been renamed to "click mode" since this is
really what they are: a way to override the treeview click mode by one
on a per-row basis.