treeview: Fix list possibly "losing" its items
Whenever a node-children is emitted, treeviews try to refresh their
content. However, set_children() half-assumed the children given were
an exact match of the ones shown in the treeview.

If not, it could lead to e.g. a list removing all its ITEMs (because the
given children were only CONTAINERs, and their absence was wrongly seen
as a removal).

We now make sure to specify which types are listed in children, and treat
things correctly.

It could be better: in the previously mentionned case, we just don't
remove anything. We could list only matching rows & removes those indeed
missing (e.g. if we have one more CONTAINER than children), but we don't
because it seems too much complication for a rare enough case.