treeview: ctxt: Fix aliases {second_,}sort_order
They should only resolve to the list of columns one can sort by, but not
be a submenu. Up to the user to put them in a submenu or not, maybe add
other items as well in there, etc

Also, order the columns in the current order instead of it being random,
and use the title (and not name) as node name/menu label.