treeview: load_arrangement(): Fix preserving columns
We used to look for a column with the same ct, and use it. This could mean
reset the ct from column foo for new column bar, while the column foo will
be added afterwards.
In which case, column foo's options would have to be reloaded instead of
preserving its current/in-memory options.

We now look for the same column, and re-use it if found. If not, we look
for a matching ct, but only one that isn't of a column that will be
preserved.