donna: parse_fl(): Add "user-parsing" (prefix, alias, etc)
Now when doing the parsing, prior to the contextual parsing, we do what
we're calling "user parsing" :

- prefixes: can be defined in "donna/prefixes/<#>" as a "prefix" and a
  "replacement" (both required string options). Prefix cannot start with
  a letter, and when matched will be replaced w/ their replacement.
  Optional boolean option is_strict means the prefix must be followed
  by something that isn't a space (to allow an alias of the same string)
  Optional boolean option is_home_dir will have a replacement of the
  user's home dir (no support for "~user" syntax (yet)).

If there was no prefix match, then we look for the first character to be
either a space, colon or slash. If it's a colon, we assume a full location
and move on. If a space (or nothing) we look for aliases, and if a slash
we try the relative path resolution.

- aliases: can be defined in "donna/aliases/<ALIAS>" as "replacement" to
  be replaced with. When there was nothing after, "replacement_no_args"
  is tried first.
- relative path resolution: if the current location (of active-list) is
  in a non-flat domain, we try to "resolve" it, i.e. remove any "./" &
  remove any "../" alongside the parent, to get a full path string (where
  multiple & trailing slashes have also been stripped).