app: {get,trigger}_node can now user-parse the FL first
In use in many places (e.g. command arguments, mark destinations, etc)
this allows to use prefixes, aliases & other relative resolution wherever
a node is needed, making things much easier/simpler.

Note that in commands, nodes & rowids expect "<XX>" to be an intref,
so we now also check the closing char. Rowids also have the "[XX,YY]"
syntax for rows.
So using '<' or '[' as prefix could lead to error (e.g. invalid intref)
in cases where the last char is a match. If that happens, add a trailing
spacen or don't use the prefix :)