provider-task: Fix possible segfault & tweak tasks deletion
We used to get the task from a node location, assuming that the address
was valid/in use, since we created the node.

However, there's a race condition that can make this false: if the node
gets invalidated right after we got the location (i.e. the task is removed
from task manager at the same time in another thread).

We now properly ensure the TM knows about the task, and take a ref on it
so we can safely interact with the task.

Also, change how tasks deletion works, specifically when there are more
than one node/task to delete, we try and delete them as we loop on sources,
and in case some failed to be deleted, the task will failed reporting all
the tasks that couldn't be deleted.

IOW, much like when deleting files, the task can fail if some tasks failed
to be deleted, while others were.