treeview: No error on {,column_}set_option if memory/config differ
We used to trigger an error when trying to set an option in CURRENT/ASK
when memory & config didn't have the same value (i.e. we couldn't confirm
that our from/current save_location was correct), IOW when there had
been changes in memory.

This was actually more annoying than anything else, even preventing to
set options or forcing to use other save_locations, etc -- so it's gone.