treeview: Refactor set_option()
We now get the list of supported options, to check that option exists
but also get its type/extra. We can then convert/check the (string
representation of the) value to set, and finally set it.

Also, for options that are extra LIST_FLAGS, it was actually only possible
to specify one infile value to be toggled from current value. Now, one can
specify a (comma-separated) list of flags, to be set.
Prefixing said list with a comma to make it a toggle from current value.