treeview/columntypes: Refactor column_set_option()
CTs can now return the list of options they support, w/ type & extra. This
is used by treeview to make sure the option is supported, and convert/check
the value, using extras if needed, much like for treeview options.

So in set_options() CTs now get a pointer to the value, and a flag toggle
for FLAGS when the value needs to be toggled from current value.
They are sure the option exist, and the value is valid (according to
extra). And of course all parsing is done the same for all treeview/column
options.

Also, ct-perms now has an extra "sort-perms" for the permissiosn sorting
criteria.