command: Fix node_trigger() being GUI (now INTERNAL)
It doesn't have to be GUI, and was only so in case we needed to set
location of active-list. This is now done via a new GUI task.

Besides, when we call cmd_node_popup_children() there was a risk of
blocking things, as this is the task worker for an INTERNAL task!