treeview: Fix handling key before last key has been fully processed
It was possible to have a key processed before the last one had been fully
processed and keys were reset. This could happen in trigger_fl() would
start a new main loop (e.g. via its get_node() in a "slow" provider) or
if the triggered node was FAST/GUI and processed events.

For example, pressing "eop" fast enough would get the "p" to be processed
during the trigger_fl() for "eo" and then be processed as "ep" resulting
in two floating windows for permission editing at once.