command: Fix handling of uint args
Commands only support ARG_TYPE_INT, used for both gint & guint arguments.
However, we shouldn't cast to guint and assume it's fine, since the
command parser deals with gint.

Cast to gint, and check it is >=0 (and therefore safe to cats to guint),
else fail the command for invalid value.