task-process: Add import_environ()
Because when we use {un,}setenv() we actually switch from using the
current environement to a brand new one, with only what we (un)set.

This might be good to create a brand new/clean environement, but when
we only want to change/add a variable, simply get import_environ() first.