Add embedded terminal support (via XEMBED)
We add a new UI element "terminal" which can be used in the layout. Much
like treeviews, a name must be given, and the related config will be found
in "terminals/<NAME>"

A terminal isn't visible by default. Starting something in an embedded
terminal is done via/as adding a new tab, as terminals handle their own
tabs automatically.

Tabs are identified via their page number (current position, starting at
1 (unlike GtkNotebook)) and IDs (unique number assigned on tab creation,
and vaid until tab is closed/removed).

XEMBED protocol is used, and the terminal command line use %w for the
window id to be used. urxvt supports this, as does xterm. Probably more.

We have a problem: clicking the terminal window doesn't focus it. This
is a PITA that doesn't seem to be easilly solved. If the terminal was
to send XEMBED_REQUEST_FOCUS, it would be awesome. But they don't, and
it seems like that's normal/expected behavior -- and our responsability,
as embedder, to do the "window managing job" of handling focus.

Meanwhile, there are commands to focus a tab, and one can double-click
on a tab to focus its terminal.