Add embedder to handle click-to-focus in terminals
When embedding a terminal (via GtkSocket) clicking inside the terminal window
doesn't focus it. This is because (most) terminals don't have a click-to-focus
policy, and as such won't send an XEMBED_REQUEST_FOCUS message.

So, in order to handle that, we use a DonnaEmbedder, a GtkSocket that will place
an invisible window in front of the plugged/terminal window, to process mouse
events - and then deliver events to subwindows as if the invisible window hadn't
been there, so everything else (e.g. wheel scroll, selection, etc) still works.

It's not perfect, but seems to do the job just fine.