Add DonnaPattern for patterns w/ many types
Now when doing string matching we have a common interface with a few extra
features.

The string can start with one of the following to specify which mode to use:

- a double quote (") for pattern mode
- a single quote (') for search mode
- a caret (^) for begin with mode
- a dollar sign ($) for end with mode
- a tilde sign (~) for case-insensitive match
- an equal sign (=) for case-sensitive match
- a greather than sign (>) for regex mode

If none are used, it'll be search mode unless at least one wildchar ('*' or '?')
was used, then it is pattern mode.

Additionally, if the string first starts with a pipre (|) then other pipes are
used as separators so more than one "pattern" can be specified.