Use DonnaPattern where we only had glob-like patterns
That is for arrangements and filtering via ct name & text.

Note that, for ct-name filtering, there is a break of compatibility: it
supported specifying multiple patterns separated via pipe signs. Now this is
still supported, but requires an extra pipe as prefix to enable it.

(Also means pipe signs can now be used as part of a pattern.)