list: Fix possible critical warning on multiple refreshes
If e.g. holding F5 for a bit to trigger multiple refreshes, we could be
cancelling a get_children_task that hadn't yet started, in which case
it was set CANCELLED right away and therefore had its callback run right
away as well.

That would clear priv->get_children_task, which we then tried to use (to
unref it). (So it was also causing a memory leak.)