task: Fix possible double-free
When cancelling a task that had never been started (i.e. going from STOPPED
or WAITING to CANCELLED directly) the task had never ran, but its callback
was still called.

As a result, on finalize we would wrongly free callback_data. Make sure
to set flag task_ran in order not to free it.