filter/ct-s: Update the API re: filter compilation
Now a DonnaFilter can be compiled ahead of time, via compile() which
allows to catch any potential (syntax) error. In order to do so,
columntype-s now have a refresh_filter_data() that can be used to set
their filter data.
The match (now done via renamed is_filter_match()) doesn't take the
(string) filter, only the filter_data.

Back in filter, when is_match() is called, is the filter isn't compiled
then it is done automatically. If it fails, a warning is emitted (as
there's no more GError) and it returns FALSE.