treeview: status: "avhsF" now support an extra w/ recursion
An extra can be specified for all those variables, which will be used:

- %F: when a VF is set
- others: the extra is splitted using commas as separators (use %, for
  literal commas), then what is shown is based on how many string were
  specified (up to 3):
With only 1 string, it will be shown unless the reference (i.e. the number
of items the variable refers to) is zero.
With 2 strings, nothing is shown if the reference is zero, the first one
is shown if it is one, else the second one is shown.
Lastly, with 3 strings, the first one is shown is reference is zero, the
second one if it's one, else the last one is shown.

And %a doesn't show nothing/first of 3 strings when it is zero, but when
it is the same as %v (i.e. all rows are visible)