filter: Fix parsing patterns starting with quote (w/out column name)
In a filter, the filter string (for the ct) can be between quotes, but
only when the column name was specified. If it wasn't, we default to "name"
and then the entire string is the filter string for name, regardless of
whether it starts with a quote or not (it could, e.g. to force-enable
pattern mode).