provider-command: Fix handling of tasks for "subcommands"
When running another commands to get their return value as argument, a
"fake" task is used and we try to run it in the current thread. Special
care must be taken when it comes to GUI, or not blocking UI thread.

This is now done by checking whether we are in the UI thread or not
(regardless of task/command visibility, since it could be FAST), and
using a new thread when needed (i.e. we're in UI but command isn't GUI
not FAST, or we're not in UI but command is GUI).