task: Emit notify signals in current/task's thread
While callback & timeout must always be in the main/UI thread, there's
no reason for such a need when it comes to notify signals. (So far those
are only used by the task manager, to "relay" the change to the node for
the task, if any. All of which can be done in any thread; and if there's
e.g. a treeview showing data (from the node), it will handle switching
to the UI thread as it always does.)

One exception: from is_cancelling() when the task is paused, since we
then own a lock on the task, and need to avoid deadlocks. (Every other
notify is done without owning a lock on the task.)

Also: this could lead to "too many files open" errors if the main thread
got flooded with get_node() tasks (e.g. search results w/ 1000's of results
per second), as even after the task ran and was used, there would remain
some idle sources for notify emission pilling up in the UI thread, all
with a ref on the task (only thing keeping it alive) and a blocking fd
open (leading to lots of eventfd, hence the error).