list: Don't connect to provider's node-children
This signal is intended for trees to update their maxi-expanded rows
when the list gets the content. It isn't much useful (if ever) for lists,
but would cause slowness since every change of location would trigger a
node-children signal, and therefore the list be updated twice.
(And when there's lots of rows, having to re-sort things might take a few
seconds.)