list: Better handle being flooded w/ node-new-child
That is, in search results when getting lots of results found very
quickly, the UI would get unresponsive as it is flooded by signals and
keeps adding nodes & resorting tree.

Now, until the model has 100 rows we just add right away. After that, we
store them in a temporary array and set up a timeout (1s), to try and let
them pile up.
That way, we can add a bunch of nodes at once and resort the tree only
once, which makes things faster/the UI more responsive.

We also only add a few nodes and process then others events, again to
try & keep the UI as responsive as possible.