provider-exec: Add setting workdir on cmdline
Now a command line (so in every mode but desktop_file) can start with
either WD= or WORKDIR= followed by the path to set as working directory.
If needed (i.e. contains spaces), the path should be quoted.

If not specified (or en empty string) then the current directory will be
used.