filter/app: Fix how getting ct_data works
app now keeps the column's ct_data with a ref_count, and when a filter
needs one it takes a ref on it. Once done, it unrefs it.

Also, app now watches config and refreshes the ct_data if needed. If it
has been ref-ed (i.e. in use by a filter) it just creates a new one, and
the old one will be free-d on last unref.

This makes sure that if the ct_data is invalidated/refreshed while a
filter uses it, it keeps its (valid) ct_data all along.

Noting that when using filter with/through a treeview, it must be in
thread UI and goes through its ct_datas. But when via app, it works on
any thread.