provider-exec: Parse environment variables
Now when creating a node, location is processed and environment
variables are resolved. $FOOBAR will be replaced with its value (if any),
quoted (to make things easier). $"FOOBAR can be used for as-is value.
Use $$ to include a literal dollar sign.

This is all done on node creation, so the node's location will then be
the resolved string (i.e. this isn't a shell-like parsing or anything.)

Done this way cause it should work for most things, without the need to
handle the entire shell-compatible command line parsing.