node: Remove new_from_node()
This was a bad idea (also, not (yet) used).

Bad idea because to make e.g. a node based on one from "fs" for listing
the same file more than once as search results, you should create a node
on your own (refing the node from fs and "replicating" its properties).

This was bad because the original refresher/setter would be called on nodes
that don't belong to the original provider, which could lead to trouble.
Because a node_peek_provider() wouldn't return the expected provider, or
because the node's location wasn't actually owned by the provider (but the
"new" one). All things that would badly fail.

Besides, the coming addition of user_data to refresher/setter for
extra properties wouldn't help things either.