node: Add visibility for refreshers
When creating a node, a visibility is given for the refresher; One is
also given when adding properties.

When the node creates a task to refresh properties, it'll use those to
determine the best/appropriate visibility for the task.

E.g. refeshing nodes in mark or config is in-memory, so FAST: no need to
create/use threads. Same goes for all properties with gtk_true as
refresher.