app: Fix handling menu destruction (re: memory leak)
We used to ref_sink/unref the menu, to work around the forced floating
ref on menu and a GTK bug that would leak on destroy.
But, with out patch this isn't needed anymore, and in fact since we didn't
call destroy on the menu, there now was a leak (since GTK still has its
strong ref on the menu).

This handles things properly in all cases, using new help popup_and_destroy()
when patched, or doing the old ref_sink/unref when not to avoid the leak.