provider-task: Make task to refresh_tm INTERNAL_FAST
INTERNAL_FAST because it should be pretty fast (it is 100% in memory) so
there's no need to need/use a(nother) thread just for that.

Also, if all (5) internal threads were to be busy waiting on public
threads (which could happen via e.g. custom properties), then there
wouldn't be a thread to refresh the task manager, so no waiting tasks
would be started, and we've hit a some kind of deadlock situation.