treeview: Avoid multiple calls to refresh node properties
If the ct asks for refeshing some properties to render, we trigger a
refresh. However, if another drawing operation was done before the
refresh completed, we would then start another task to refresh.

We now check if we've already a task pending to refresh the same properties
for that same node, and if so do nothing.