task: Add visibility INTERNAL_LOOP
This visibility is used with a "pre-worker". The idea is that when calling
donna_app_run_task() it will if needed create a new thread and start a
new main loop in it, then attach an idle source that will pre-run the
task.

Then, the pre-worker could return TRUE to have the task run right away,
or it could e.g. attach a timeout source (that will run the task). This
allows e.g. to have a few tasks to "pile up" in the main loop, and then
have one script running for all tasks at once (this is all very custom
properties talk).