node: Fix possible multithread condition in node_refresh()
In the worker when refreshing properties, we connect to the "notify"
signal to keep track of which properties are refreshed. Although any
refresh from our called refresher will happen in this thread, there's
no guarantee that another thread won't refresh one of those properties
on the node at the same time.

So we need a mutex around refreshed to handle such cases.