treeview: Minor optimization
Or fix. refresh_cb() is a task callback, and as such always runs in the
thread UI, so there's no need for a mutex.

(We could connect to notify::state waiting for tasks to go POST_RUN, in
which case the signal handler would run in the task's thread, but there's
no point really in adding such extra complications.)