node: Add functions to get array of tasks to refresh properties
To take full advantages of VISIBILITY_LOOP on refreshers, it might be
good to get an array of tasks: the node's internal one (if needed, to
call refreshers directly), and those from refresher_task (i.e. w/ said
VISIBILITY_LOOP), so the node's internal task doesn't need to be there
for no reason, nor won't block waiting for other refreshers.