app: custom props: Fix refresher not really being blocking
The only way the custom refresher should ever be called is for blocking
calls (e.g. donna_node_get() blockingly), and so it is meant to block.

Since we need to wait for the task-process, we need to do so blockingly
as well, instead of using donna_task_wait_for_it() which could start a
new main loop (to avoid blocking).

Example: go in folder w/ a custom property, and set a VF using said CP.
If not all rows were visible, and non-visible rows haven't had their CP
refreshed, the filter (via donna_filter_is_match()) would trigger the
blocking refresh call.

If we were to start a main loop during that time, the UI does not block,
but the main loop processes the notify signal, refilters the node, thus
making the pending refiltering (which called the filter) work on a now
free-d memory... and (likely) segfaults.