treeview: Don't refresh props for non-visible rows
When a ct gives a list of props to refresh in order to be able to render
the cell, we'll ignore non-visible rows.

This is because otherwise if there was e.g. 2 custom props (CP) on the
same nodes, 1 of them being preload, then the node-updated would trigger
a row-updated and have the treeview redraw the row, even when not visible.
And we would then trigger refreshes for the other CP, when that might not
be wanted.