ct-perms: Fix get_props()
The analysis of the format string to determine which properties was
actually needed/used was bugged, e.g. not accounting for mode when showing
a colored use/group name.

But, get_props() should return all properties that might be used for render,
sorting or filtering. So it shouldn't be based on analyzing the format
string, the ctdata is to be used when the property is defined there.

(This will be even more important with custom properties and ON_DEMAND
property refreshing.)