ct-time: Add new specifier %f for "fluid" format
That is, for today only the time is shown; For yesterday time is prefixed
with "Yesterday"; Within the last 7 days time is prefixed with weekday
name; Else date & time are shown.

Options allows to specify the format for time & date, and whether to use
abbreviated weekday names or not.

Also make "%f" the new default format.