app: socket: Add commands TRIGGER and CANCEL
TRIGGER allows to trigger any node, i.e. it gives full power to everything
in donna. The given full location is parsed as expected, so prefixes/aliases
are supported.

On success, an ID is given (e.g. "OK TRIGGER 1") which can then be used
to cancel the task, should one want to.
It will also be used when the task reaches POST_RUN state, as either:
- "DONE <ID>[ <RETURN VALUE>]"
- "CANCELLED <ID>"
- "FAILED <ID>[ <ERROR MESSAGE>]

The return value will be as intref when it was a treeview, node, terminal,
or an array (or nodes or strings).

Scripts should free the intrefs they get once done, we just need to add
a command for that :)