task: Remove VSIBILITY_LOOP; Add pre-worker support
Instead of having a dedicated visibility which required a dedicated
thread/main loop and was mostly hackish (in that task workers were
reduced to close to no-op as the actual work had to be done before/in
other tasks, else it'd block the dedicated main loop), now every task
can have a pre-worker set.

When that happens, the task shall be prerun first, so the preworker can
do its work, maybe installing sources in the main loop to get there, then
set the task preran and trigger its actual start, which will then happen
as regular, so in a dedicated/UI thread as per its visibility.

Custom properties still "hack" this since the refreshing tasks are just
waiting (via sources from preworker) for the task-process to run, due to
the whole buffering/groupping that happens there.

But, it will be useful for all commands, so the parsing of argument can
be done via preworker, leaving the worker be only the actual command.
It's better/more obvious, and also removes any need for recursive main
loops started for "subcommands" when parsing arguments and possible
issues related to this.