command: Use pre-worker to parse arguments
Instead of having to resort to (recursive) main loops to handle running
subcommands as arguments, which was bas design, parsing arguments is now
done in the task's pre-worker.

If there's a subcommand to run, the corresponding task is created and a
source added to the main loop, to resume parsing. Once ready, the actual
command is started, leaving the worker only be the actual function of the
command, and nothing else.