Add provider filter to handle filters
Filters are now handles via a provider, which means they aren't used as
strings but nodes via commands. (Behind the scenes, the same DonnaFilter
objects are used.)

The main benefit from this, besides making it more in line with how donna
works, is that now filters can be defined in config, to assign a name,
alias and/or icon name. So when filters are presented in a menu, they can
have a name (e.g. "From yesterday") and icon. It is also possible to simply
use the alias to reference the filter.

Specifically, nodes in filter will always have for location the filter
string, but one can get the node by using either said filter string (e.g.
"filter:size>100M") or the alias, prexifed by a pipe (e.g. "filter:|large")

This applies not only in commands, but also in other part of the config
where filters are used, e.g. color filters or custom properties. There,
a filter string can be used, or simply an alias prefixed with a pipe (i.e.
a valid location in filter).

Because now nodes are used when it comes to filter, a special node with
an empty string as location ("filter:") is used to mean "no filter".
Point is, when using e.g. tv_get_visual_filter() a node needs to be
returned, so when no filter is set this node will be returned. (It could
also be used to unset any VF, though the argument remains optional and can
simply be omitted still.)

Because it is a node in "filter" one can still assign a name, alias and/or
icon to it, as for any filter.