command: Add node_set_property()
Can set property that are string, int, uint or date. This is pretty basic
since you need to know beforehand the property's type and there's no
things like "now" or "an hour ago" and the likes, but that can be done
via a treeview's column (e.g. tv_column_set_value())