app: Implement our own lock to separate things
Instead of one RWLock for visuals & providers and a (recursive) mutex for
everything else, we now use our own locking mechanism (mutex+cond+flags)
so we can separate everything (as if everything had its own lock) and
still use RW for visuals & providers (though, again, separatedly).

Much better overall and helps prevent situations where the lock is taken
for one reason and then blocking (possibly deadlocking) for another
reason.

One special case: when having a lock on col_ct_datas and triggering a
refresh_ct() then option_cb() can be triggered, thus causing a deadlock.
We handle this by failing to take the lock (instead of blocking for it)
and using an idle source to process the option_cb later (once the lock
is available).