contextmenu: Add TYPE_COMBINED for trigger+container
A new context-type "combined" allows to define a user item with the
usual option "trigger" for the trigger part, but also option "container"
for the container part.

This is obviously intended to be used w/ option "submenus" set to "combine"
since the resulting node will be a container that can also be triggered.

This allows to define the trigger (and other props, e.g. name, icon...)
as usual, all the while also defining the submenu's content (other than
by making use of the <...> syntax for other items).