Add provider mru
An MRU list can contain either nodes or strings. In the former case, its
node will be a container, else it'll be an item (and in that case, the
only way to get items is via command mru_get_strings).

Not unlike filters, asking for a node mru:foobar will automatically create
said MRU (foobar, as MRU of nodes) if it doesn't exist already. So there's
nothing to do to create one, though command mru_new allows to do creation
only (i.e. fails if exists) and, of course, to create an MRU of strings.