scripts: Add sel_filter
An example of what can be done in scripts using donna-trigger.

Here when clicking the "Selection Filter" menuitem, we get the last
filter used via MRU sel_filter to use it as default (its alias if any,
else the actual filter).

Then the filter is added to the MRU before being applied to select
matching nodes.